+++ package.json
@@ -19,9 +19,6 @@
     },
     "devDependencies": {
         "@tsconfig/node18": "^1.0.1",
-        "@types/cookies": "^0.7.7",
-        "@types/express-session": "^1.17.6",
-        "@types/mock-req-res": "^1.1.3",
         "@typescript-eslint/eslint-plugin": "^5.52.0",
         "@typescript-eslint/parser": "^5.52.0",
         "aws-cdk-lib": "2.62.2",
@@ -32,7 +29,6 @@
         "eslint-plugin-import": "^2.27.5",
         "eslint-plugin-prettier": "^4.2.1",
         "husky": "^8.0.0",
-        "mock-req-res": "^1.2.1",
         "prettier": "^2.8.4",
         "sst": "^2.0.18",
         "typescript": "^4.9.5"
@@ -47,10 +43,5 @@
         "site/*",
         "site",
         "shared-ts/*"
-    ],
-    "dependencies": {
-        "cookies": "^0.8.0",
-        "nookies": "^2.5.2",
-        "winston": "^3.8.2"
-    }
+    ]
 }
