+++ site/components/layout/CookieBanner.test.tsx
@@ -1,11 +1,8 @@
 import renderer from "react-test-renderer";
-import Cookies from "universal-cookie";
-import { vi, describe, it, expect } from "vitest";
+import { describe, it, expect } from "vitest";
 import CookieBanner from "../layout/CookieBanner";
 
 describe("CookieBanner", () => {
-    Cookies.prototype.set = vi.fn();
-
     it("should render correctly", () => {
         const tree = renderer.create(<CookieBanner />).toJSON();
         expect(tree).toMatchSnapshot();
