+++ site/package.json
@@ -22,14 +22,17 @@
         "@types/node": "18.11.18",
         "@types/react": "18.0.26",
         "@types/react-dom": "18.0.10",
+        "cookies": "^0.8.0",
+        "cookies-next": "^2.1.1",
         "dayjs": "^1.11.7",
         "edge-csrf": "^1.0.3-rc1",
         "govuk-frontend": "^4.4.1",
+        "mock-req-res": "^1.2.1",
         "next": "13.1.2",
+        "nookies": "^2.5.2",
         "react": "18.2.0",
         "react-dom": "18.2.0",
         "typescript": "4.9.4",
-        "universal-cookie": "^4.0.4",
         "uuid": "^9.0.0"
     },
     "devDependencies": {
@@ -38,6 +41,8 @@
         "@babel/preset-react": "^7.18.6",
         "@babel/preset-typescript": "^7.21.0",
         "@testing-library/react": "^14.0.0",
+        "@types/cookies": "^0.7.7",
+        "@types/mock-req-res": "^1.1.3",
         "@types/react-test-renderer": "^18.0.0",
         "@types/uuid": "^9.0.0",
         "@vitejs/plugin-react": "^3.1.0",
@@ -53,6 +58,7 @@
         "sass": "^1.57.1",
         "sst": "^2.0.18",
         "tailwindcss": "^3.2.4",
-        "vitest": "^0.29.1"
+        "vitest": "^0.29.1",
+        "winston": "^3.8.2"
     }
 }
