+++ site/pages/cookies.test.tsx
@@ -2,7 +2,7 @@ import renderer from "react-test-renderer";
 import { vi, describe, it, expect, afterEach } from "vitest";
 import Cookies, { CookiePreferencesProps, getServerSideProps } from "./cookies";
 import { CookiePolicy } from "../interfaces";
-import { getMockContext } from "../pages/testData/mockData";
+import { getMockContext } from "../testData/mockData";
 
 describe("pages", () => {
     describe("cookies", () => {
