+++ site/pages/cookies.tsx
@@ -18,7 +18,7 @@ export interface CookiePreferencesProps {
 const Cookies = ({ settingsSaved, trackingDefaultValue, csrfToken }: CookiePreferencesProps): ReactElement => (
     <TwoThirdsLayout title={title} description={description} hideCookieBanner>
         {settingsSaved ? (
-            <div className="information_box__success block p-[20px] mb-[50px] border-4 border-govGreen">
+            <div className="information_box__success block p-5 mb-12.5 border-4 border-govGreen">
                 <h2 className="govuk-heading-m">Your cookie settings were saved</h2>
                 <p className="govuk-body">Return to this page at any point to change your cookie settings.</p>
                 <p className="govuk-body">
